chore: hooks, auth and utils directory changes
