+++ app/auth/constants.ts
@@ -1,0 +1,1 @@
+export const USER_SESSION_KEY = 'userId';
