+++ app/auth/cookiesSessionStorage.ts
@@ -1,0 +1,15 @@
+import { createCookieSessionStorage } from '@remix-run/node';
+import invariant from 'tiny-invariant';
+
+invariant(process.env.SESSION_SECRET, 'SESSION_SECRET must be set');
+
+export const cookieSessionStorage = createCookieSessionStorage({
+  cookie: {
+    name: '__session',
+    httpOnly: true,
+    path: '/',
+    sameSite: 'lax',
+    secrets: [process.env.SESSION_SECRET],
+    secure: process.env.NODE_ENV === 'production',
+  },
+});
