+++ app/auth/getUserId.ts
@@ -1,0 +1,8 @@
+import { USER_SESSION_KEY } from './constants';
+import { getSession } from './getSession';
+
+export async function getUserId(request: Request): Promise<string | undefined> {
+  const { session } = await getSession(request);
+  const userId = session.get(USER_SESSION_KEY);
+  return userId;
+}
