+++ app/auth/redirectUserId.ts
@@ -1,0 +1,14 @@
+import { redirect } from '@remix-run/node';
+import { getUserId } from './getUserId';
+
+export async function requireUserId(
+  request: Request,
+  redirectTo: string = new URL(request.url).pathname,
+): Promise<string> {
+  const userId = await getUserId(request);
+  if (!userId) {
+    const searchParams = new URLSearchParams([['redirectTo', redirectTo]]);
+    throw redirect(`/login?${searchParams}`);
+  }
+  return userId;
+}
