+++ app/routes/index.tsx
@@ -1,5 +1,5 @@
 import { Link } from '@remix-run/react';
-import { useOptionalUser } from '~/utils';
+import { useOptionalUser } from '~/hooks/useOptionalUser';
 
 export default function Index() {
   const user = useOptionalUser();
