+++ app/utils/isUser.ts
@@ -1,0 +1,5 @@
+import type { User } from '~/prisma-actions/user.server';
+
+export const isUser = (user: User) => {
+  return user && typeof user === 'object' && typeof user.email === 'string';
+};
