+++ app/utils/validateEmail.ts
@@ -1,0 +1,3 @@
+export const validateEmail = (email: string) => {
+  return typeof email === 'string' && email.length > 3 && email.includes('@');
+};
